% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-causamed.R
\docType{package}
\name{causamed}
\alias{causamed}
\alias{causamed-package}
\title{Causamed: A package for hopeless mediation things.}
\description{
Causamed: A package for hopeless mediation things.
}

