% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fmatch}
\alias{fmatch}
\title{utils}
\usage{
fmatch(v, A, l = NULL, mat = FALSE)
}
\arguments{
\item{v}{The matrix from which elements are extracted}

\item{A}{The vector indexing the elements of the matrix v}

\item{l}{The length of A, given to speed up slightly}

\item{mat}{indicator that the indexing vector should be returned instead of of the indexed matrix}
}
\description{
Functions used to run the package
}

