% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_iptw.R
\name{med_iptw.mkdat}
\alias{med_iptw.mkdat}
\title{med_iptw}
\usage{
med_iptw.mkdat(orig_dat, A, M, Y, C = "", L = "")
}
\arguments{
\item{orig_dat}{The original dataset}

\item{A}{Text string containing the name of the exposure variable}

\item{M}{Text string containing the name of the mediator variable}

\item{C}{vector of text strings containing the name or expression of confounder variables}

\item{L}{vector of text strings containing the name of exposure-induced confounders of M -> Y}
}
\description{
Computes CDE(M) for a given mediator level, in a setting with an exposure-induced confounder of
the mediator-outcome association
}

