% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_iptw.R
\name{med_iptw}
\alias{med_iptw}
\title{\code{med_iptw} Computes CDE(M) for a given mediator level, in a setting with an exposure-induced confounder of
the mediator-outcome association}
\usage{
med_iptw(dat, A, M, Y, C = "", L = "", regtype = "gaussian",
  noint = FALSE, boot = 100, quants = c(0.025, 0.5, 0.975), mlvl = NULL)
}
\arguments{
\item{A}{Text string containing the name of the exposure variable}

\item{M}{Text string containing the name of the mediator variable}

\item{C}{vector of text strings containing the name or expression of confounder variables}

\item{L}{vector of text strings containing the name of exposure-induced confounders of M -> Y}

\item{orig_dat}{The original dataset}
}
\value{
A list containing information about the models, the controlled direct effects computed at
each specified mediator level, the total effect, and the difference between those two measures at each
mediator level
}
\description{
\code{med_iptw} Computes CDE(M) for a given mediator level, in a setting with an exposure-induced confounder of
the mediator-outcome association
}

