% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_multi.R
\name{med_multi}
\alias{med_multi}
\title{med_multi}
\usage{
med_multi(dat, X, Y, M, C, fam = gaussian(link = "identity"), boot = 10,
  nmin = 10)
}
\arguments{
\item{dat}{a dataframe containing the exposure, outcome, mediators, and confounders}

\item{X}{the exposure of interest. Currently must be categorical (or binary)}

\item{Y}{the outcome, currently must be continuous, ordinal or binary}

\item{C}{confounders of either X -> M and/or M -> Y.}

\item{fam}{specifies GLM link function and distribution of residuals. Default is gaussian(link = identity)}

\item{boot}{number of bootstrap samples used to build the 95p confidence intervals}

\item{nmin}{number of participants all categories of exposure must have; samples will be redrawn if this criterion is not met}
}
\description{
Decomposes the effect of an exposure on an outcome into a direct effect operating through a set of intermediary
variables, and a direct effect involving other pathways. A key assumption is the dual lack of confounding of the
exposure-outcome, and the mediator-outcome associations. Described in chapter 5.2.1 of Tyler's book.
}
\examples{
\donttest{my_list <- med_multi(dat = df,
X = my_exposure, M = mediator_1 + mediator_2 + ... + mediator_1*mediator_2 + mediator_1*exposure,
Y = my_continuous_outcome,
C = a_confounder + another_confounder + my_cubic_spline_term,
fam = gaussian(link = "identity"), boot = 1000)}
}

