% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_reg.R
\name{med_reg}
\alias{med_reg}
\title{\code{med_reg} Gives controlled direct effect, natural indirect effect, natural direct effect using regression-based formulas
affording the best statistical power. The procedure is described in chapter 3.1.1 of Tyler's book}
\usage{
med_reg(df, X, M, Y, C = "", noint = FALSE, mlvl = NULL, delta = FALSE,
  ref = 0, treat = 1)
}
\arguments{
\item{df}{Data frame}

\item{X}{exposure, can be continuous or categorical}

\item{M}{mediator, must be continuous for the moment}

\item{Y}{outcome, must be continuous for the moment}

\item{C}{confounders of exposure-outcome or mediator-outcome association. Can take any form}

\item{noint}{supress exposure-mediator interaction term}

\item{mlvl}{vector of values at which to compute a controlled direct effect}

\item{delta}{whether the delta method should be used to get confidence intervals}

\item{ref}{referent level of exposure, 0 by default}

\item{treat}{treatment level of exposure, 1 by default}
}
\value{
a list composed of the natural direct effect, natural indirect effect, total effect, and a controlled
direct effect for each line given in \code{mlvl = }
}
\description{
\code{med_reg} Gives controlled direct effect, natural indirect effect, natural direct effect using regression-based formulas
affording the best statistical power. The procedure is described in chapter 3.1.1 of Tyler's book
}
\examples{
\donttest{
my_list <-med_reg(my_df,
X = "my_exposure",
M = "my_continuous_mediator",
Y = "my_continuous_outcome",
C = c("confounder_1", "confounder_2"),
treat = high_exposure_level,
ref = low_exposure_level,
mlvl = seq(25, 40, by=5) ## possible BMI values for the mediator
}
}

