% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_rint.R
\name{med_rint}
\alias{med_rint}
\title{Random interventional analogue effects}
\usage{
med_rint(dat, A, M, Y, C = "", L, astar = "astar", boot = 10,
  quants = c(0.025, 0.5, 0.975), alex = FALSE)
}
\arguments{
\item{A}{Text string containing the name of the exposure variable}

\item{M}{Text string containing the name of the mediator variable}

\item{Y}{vector of text strings containing the name of the outcome variable}

\item{C}{vector of text strings containing the name or expression of confounder variables}

\item{astar}{optional alternative name for (pseudo) randomized treatment a*}

\item{boot}{number of bootstrap samples (default 10 for testing)}

\item{quants}{optionally specify own confidence interval (by default: 95p CI)}

\item{alex}{keeping this in case num specification was right}

\item{orig_dat}{The original dataset}
}
\value{
A large list containing meta-data on the procedure, along with the median and 95p confidence interval
for the randomized interventional analogue direct and indirect effects
}
\description{
\code{med_rint} computes NDEr, NIEr and TEr, the random interventional
analogues of the natural direct effect, natural indirect effect and total effect,
in the presence of exposure-induced confounding of M -> Y. Gives bootstrapped confidence
interval. To do: extend to multiple confounders. Might or might not work at the moment.
}

