% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_smean.R
\name{med_smean}
\alias{med_smean}
\title{\code{med_smean}}
\usage{
med_smean(dat, A, Y, M, C = NULL, L = NULL, boot = 10, nmin = 10,
  mlvl = NULL)
}
\arguments{
\item{dat}{a dataframe containing the exposure, outcome, mediators, and confounders}

\item{A}{the exposure of interest. Can take any form}

\item{Y}{the outcome, currently must be continuous}

\item{M}{the mediators of interest}

\item{C}{confounders of either X -> M and/or M -> Y. Can take any form}

\item{L}{the exposure-induced confounders of the association of M with Y}

\item{boot}{the number of bootstrap samples used to build confidence intervals}

\item{mlvl}{the levels of M to calculate corresponding CDE's to. Default is sample average.}
}
\value{
An S3 object of class \code{cmed_smean} containing:

k2  the coefficient of M in regression of A, M, L and C on Y

k3  the coefficient of A*M from regression of A, M, L and C on Y

g1  the coefficient g1 from the regression of A and C on partial residuals of Y

ymod1  the model of Y for the last bootstrap sample

ymod2  the model of Y residuals for the last bootstrap sample

cde  array story controlled direct effects from bootstrapped results
}
\description{
Returns the controlled direct effect CDE(M) of an exposure A on an outcome Y, not
operating through a mediator M, in the case of exposure-induced confounding of
the M->Y association. Follows section 5.3.5 of Vanderweele's book on causal
mediation analysis.
}
\examples{
\donttest{
my_list <- med_smean(data, A, Y, M, c1 + c2 + c3 + c2*c3, mlvl = quantiles(M, probs = c(0.25, 0.5, 0.75)))
}
}

