% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_rint.R
\name{rint_med.boot}
\alias{rint_med.boot}
\title{rint_med.boot: Random interventional analogue effects}
\usage{
rint_med.boot(dat, A, M, Y, C = "", L, astar = "astar", boot = 10,
  quants = c(0.025, 0.5, 0.975), alex = FALSE)
}
\arguments{
\item{A}{Text string containing the name of the exposure variable}

\item{M}{Text string containing the name of the mediator variable}

\item{Y}{vector of text strings containing the name of the outcome variable}

\item{C}{vector of text strings containing the name or expression of confounder variables}

\item{astar}{optional alternative name for (pseudo) randomized treatment a*}

\item{boot}{number of bootstrap samples (default 10 for testing)}

\item{quants}{optionally specify own confidence interval (by default: 95p CI)}

\item{alex}{keeping this in case num specification was right}

\item{orig_dat}{The original dataset}
}
\description{
Computes NDEr, NIEr and TEr, the random interventional
analogues of the natural direct effect, natural indirect effect and total effect,
in the presence of exposure-induced confounding of M -> Y. Gives bootstrapped confidence
interval
}

