% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_rint.R
\name{rint_med.decompose}
\alias{rint_med.decompose}
\title{rint_med.decompose: Random interventional analogue effects}
\usage{
rint_med.decompose(dat, Y, X, astar = "astar")
}
\arguments{
\item{Y}{vector of text strings containing the name of the outcome variable}

\item{X}{Text string containing the name of the exposure variable}

\item{astar}{optional alternative name for (pseudo) randomized treatment a*}

\item{orig_dat}{The original dataset}
}
\description{
Computes NDEr, NIEr and TEr, the random interventional
analogues of the natural direct effect, natural indirect effect and total effect,
in the presence of exposure-induced confounding of M -> Y
}
\details{
to do: extend to multiple confounders
}

