% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_rint.R
\name{rint_med.mkdata}
\alias{rint_med.mkdata}
\title{rint_med.mkdata}
\usage{
rint_med.mkdata(orig_dat, A, M, Y, C = "", L)
}
\description{
Creates the analytic dataset to compute NDEr, NIEr and TEr, the random interventional
analogues of the natural direct effect, natural indirect effect and total effect,
in the presence of exposure-induced confounding of M -> Y
}

