% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_reg.R
\name{twoway.cont.cont}
\alias{twoway.cont.cont}
\title{Med_reg}
\usage{
twoway.cont.cont(df, X, M, Y, C = "", noint = FALSE, mlvl = NULL,
  delta = FALSE)
}
\arguments{
\item{df}{Data frame}

\item{X}{exposure}

\item{M}{mediator}

\item{Y}{outcome}

\item{C}{confounders of exposure-outcome or mediator-outcome association}

\item{noint}{supress exposure-mediator interaction term}

\item{mlvl}{vector of values at which to compute a controlled direct effect}

\item{delta}{whether the delta method should be used to get confidence intervals}
}
\description{
Gives controlled direct effect, natural indirect effect, natural direct effect
}

