% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_iptw.R
\name{med_iptw}
\alias{med_iptw}
\title{\code{med_iptw} Computes CDE(M) for given mediator levels, in a setting with an exposure-induced confounder of
the mediator-outcome association. Described in chapter 5.3.1 of Tyler's book}
\usage{
med_iptw(dat, A, M, Y, C = "", L = "", regtype = "gaussian", boot = 10,
  quants = c(0.025, 0.5, 0.975), mlvl = NULL, link = logit)
}
\arguments{
\item{dat}{The original dataset}

\item{A}{the exposure of interest. Must be binary or categorical}

\item{M}{the mediators of interest. Must be binary or categorical}

\item{Y}{the outcome, currently must be continuous}

\item{C}{confounders of either X -> M and/or M -> Y. Can take any form, specified as formula}

\item{L}{the exposure-induced confounders of the association of M with Y. Can take any form}

\item{boot}{specifies the number of bootstrap samples drawn to make the confidence intervals}

\item{quants}{an optional vector of quantiles for the confidence interval (95 percent by default)}

\item{mlvl}{a matrix or table of probability-mass functions for the mediator, to calculate CDE(M). By default, mlvl is set to
the observed sample distributions}
}
\value{
An S3 object of class \code{cmed.ipw} containing:

w  the ipw used in the marginal strucutral model

cde.int  an array where cde.int[i,,] indexes a matrix corresponding to the CDE calculated for a PMF of M given in mlvl,
with each row a bootsrap replicate

cde.noint a matrix of cde given no M*A interaction, with each row a bootstrap replicate

te a matrix of total effects, with each row a bootstrap replicate

ymod1 the marginal strucutral model of Y given no interaction

ymod2 the marginal strucutral model of Y allowing a A*M interaction
}
\description{
\code{med_iptw} Computes CDE(M) for given mediator levels, in a setting with an exposure-induced confounder of
the mediator-outcome association. Described in chapter 5.3.1 of Tyler's book
}
\examples{
\donttest{
my_list <- med_iptw(dat = df,  A = my_exposure, Y = my_outcome, M = my_mediator, C = a_confounder + another_confounder, boot = 1000)
}
}

