% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_reg.R
\name{med_reg}
\alias{med_reg}
\title{\code{med_reg} Gives controlled direct effect, natural indirect effect, natural direct effect using regression-based formulas
affording the best statistical power}
\usage{
med_reg(df, X, M, Y, C = "", noint = FALSE, mlvl = NULL, delta = FALSE,
  ref = 0, treat = 1)
}
\arguments{
\item{df}{Data frame}

\item{X}{exposure, can be continuous or categorical}

\item{M}{mediator, must be continuous for the moment}

\item{Y}{outcome, must be continuous for the moment}

\item{C}{confounders of exposure-outcome or mediator-outcome association. Can take any form}

\item{noint}{supress exposure-mediator interaction term}

\item{mlvl}{vector of values at which to compute a controlled direct effect}

\item{delta}{whether the delta method should be used to get confidence intervals}

\item{ref}{referent level of exposure, 0 by default}

\item{treat}{treatment level of exposure, 1 by default}
}
\description{
\code{med_reg} Gives controlled direct effect, natural indirect effect, natural direct effect using regression-based formulas
affording the best statistical power
}

