% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_reg.R
\name{med_reg}
\alias{med_reg}
\title{med_reg}
\usage{
med_reg(df, X, M, Y, C = "", noint = FALSE, mlvl = NULL, delta = FALSE,
  ref = 0, treat = 1)
}
\arguments{
\item{df}{Data frame}

\item{X}{exposure}

\item{M}{mediator}

\item{Y}{outcome}

\item{C}{confounders of exposure-outcome or mediator-outcome association}

\item{noint}{supress exposure-mediator interaction term}

\item{mlvl}{vector of values at which to compute a controlled direct effect}

\item{delta}{whether the delta method should be used to get confidence intervals}

\item{ref}{referent level of exposure, 0 by default}

\item{treat}{treatment level of exposure, 1 by default}
}
\description{
Gives controlled direct effect, natural indirect effect, natural direct effect
}

