% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_rint.R
\name{med_rint}
\alias{med_rint}
\title{Random interventional analogue effects}
\usage{
med_rint(dat, A, M, Y, C = "", L, astar = "astar", boot = 10,
  quants = c(0.025, 0.5, 0.975), alex = FALSE)
}
\arguments{
\item{A}{a character string containing the name of the exposure variable. A must be categorical/binary}

\item{M}{a character string containing the name of the mediator variable. M must be categorical/binary}

\item{Y}{a character string containing the name of the outcome variable. At the moment, Y must be continuous}

\item{C}{list of character strings containing the name or expression of confounder variables}

\item{astar}{optional alternative name for (pseudo) randomized treatment level a*}

\item{boot}{number of bootstrap samples (default 10 for testing)}

\item{quants}{optionally specify own confidence interval (by default: 95p CI)}

\item{orig_dat}{The original dataset}
}
\value{
A large list containing meta-data on the procedure, along with the median and 95p confidence interval
for the randomized interventional analogue direct and indirect effects
}
\description{
\code{med_rint} computes NDEr, NIEr and TEr, the random interventional
analogues of the natural direct effect, natural indirect effect and total effect,
in the presence of exposure-induced confounding of M -> Y. Gives bootstrapped confidence
interval. To do: extend to multiple confounders. Might or might not work at the moment.
}
\details{
The procedure is described in chapter 5.4.2 of Tyler's book
}
\examples{
\donttest{my_list <- med_rint(dat = df,
X = "my_exposure",
M = "my_mediator",
Y = "my_binary_outcome",
C = c("a_confounder", "another_confounder"),
fam = binomial(logit), boot = 1000)}
}

