% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_rint.R
\name{rint_med.decompose}
\alias{rint_med.decompose}
\title{rint_med.decomposes}
\usage{
rint_med.decompose(dat, Y, A, astar = "astar")
}
\description{
part of the functio calculating CDE(M) at different values or distributions of M, and computes weights used in
master function
}

