% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_rint.R
\name{rint_med.mkdata}
\alias{rint_med.mkdata}
\title{med_rint: Random interventional analogue effects
rint_med.mkdata Creates the analytic dataset to compute NDEr, NIEr and TEr, the random interventional
analogues of the natural direct effect, natural indirect effect and total effect,
in the presence of exposure-induced confounding of M -> Y}
\usage{
rint_med.mkdata(orig_dat, X, M, Y, C = "", L)
}
\arguments{
\item{orig_dat}{The original dataset}

\item{X}{Text string containing the name of the exposure variable}

\item{M}{Text string containing the name of the mediator variable}

\item{Y}{vector of text strings containing the name of the outcome variable}

\item{C}{vector of text strings containing the name or expression of confounder variables}

\item{L}{vector of text strings containing the name of exposure-induced confounder of M -> Y}
}
\description{
to do: extend to multiple confounders
}

